orquestador-service-0/requirements.txt
fastapi
uvicorn
google-cloud-storage
python-multipart
requests
python-dotenv
sqlalchemy
google-cloud
